Resolve https://github.com/raedahgroup/dcrios/pull/526#pullrequestreview-311311285
- refactor sync details view to subscribe to past events
- refactor syncmanager to update number of connected peers on different sync actions
