+++ Decred Wallet/Extensions/UIStackView.swift
@@ -10,7 +10,7 @@ import UIKit
 
 extension UIStackView {
     // Embed the UIstackiew components in a parent UIview with rounded edges as UIstackview does disregards it's layer.cornerRadius directive
-    func cornerRadius(_ radius: CGFloat) {
+    func setCornerRadius(_ radius: CGFloat) {
         // Check if the stackview already has a corner radius providing view embedded
         if self.subviews[0].accessibilityIdentifier == "cornerRadius" {
             return
